refactor (#301)
