+++ git-worktree/src/index/checkout.rs
@@ -1,103 +1,4 @@
-use crate::fs;
 use bstr::BString;
-
-/// A cache for directory creation to reduce the amount of stat calls when creating
-/// directories safely, that is without following symlinks that might be on the way.
-///
-/// As a special case, it offers a 'prefix' which (by itself) is assumed to exist and may contain symlinks.
-/// Everything past that prefix boundary must not contain a symlink. We do this by allowing any input path.
-///
-/// Another added benefit is its ability to store the path of full path of the entry to which leading directories
-/// are to be created to avoid allocating memory.
-///
-/// For this to work, it remembers the last 'good' path to a directory and assumes that all components of it
-/// are still valid, too.
-/// As directories are created, the cache will be adjusted to reflect the latest seen directory.
-///
-/// The caching is only useful if consecutive calls to create a directory are using a sorted list of entries.
-#[allow(unused)]
-pub struct PathCache {
-    stack: fs::Stack,
-    /// If there is a symlink or a file in our path, try to unlink it before creating the directory.
-    pub unlink_on_collision: bool,
-
-    /// just for testing
-    #[cfg(debug_assertions)]
-    pub test_mkdir_calls: usize,
-}
-
-mod cache {
-    use std::path::{Path, PathBuf};
-
-    use super::PathCache;
-    use crate::{fs, os};
-
-    impl PathCache {
-        /// Create a new instance with `root` being the base for all future paths we handle, assuming it to be valid which includes
-        /// symbolic links to be included in it as well.
-        pub fn new(root: impl Into<PathBuf>) -> Self {
-            let root = root.into();
-            PathCache {
-                stack: fs::Stack::new(root),
-                #[cfg(debug_assertions)]
-                test_mkdir_calls: 0,
-                unlink_on_collision: false,
-            }
-        }
-
-        /// Append the `relative` path to the root directory the cache contains and efficiently create leading directories
-        /// unless `mode` indicates `relative` points to a directory itself in which case the entire resulting path is created as directory.
-        ///
-        /// The full path to `relative` will be returned for use on the file system.
-        pub fn append_relative_path_assure_leading_dir(
-            &mut self,
-            relative: impl AsRef<Path>,
-            mode: git_index::entry::Mode,
-        ) -> std::io::Result<&Path> {
-            #[cfg(debug_assertions)]
-            let mkdir_calls = &mut self.test_mkdir_calls;
-            let unlink_on_collision = self.unlink_on_collision;
-            self.stack.make_relative_path_current(
-                relative,
-                |components, stack: &fs::Stack| {
-                    let target_is_dir = mode == git_index::entry::Mode::COMMIT || mode == git_index::entry::Mode::DIR;
-                    if components.peek().is_some() || target_is_dir {
-                        #[cfg(debug_assertions)]
-                        {
-                            *mkdir_calls += 1;
-                        }
-                        match std::fs::create_dir(stack.current()) {
-                            Ok(()) => {}
-                            Err(err) if err.kind() == std::io::ErrorKind::AlreadyExists => {
-                                let meta = stack.current().symlink_metadata()?;
-                                if !meta.is_dir() {
-                                    if unlink_on_collision {
-                                        if meta.is_symlink() {
-                                            os::remove_symlink(stack.current())?;
-                                        } else {
-                                            std::fs::remove_file(stack.current())?;
-                                        }
-                                        #[cfg(debug_assertions)]
-                                        {
-                                            *mkdir_calls += 1;
-                                        }
-                                        std::fs::create_dir(stack.current())?;
-                                    } else {
-                                        return Err(err);
-                                    }
-                                }
-                            }
-                            Err(err) => return Err(err),
-                        }
-                    }
-                    Ok(())
-                },
-                |_| {},
-            )?;
-            Ok(self.stack.current())
-        }
-    }
-}
 
 #[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
 pub struct Collision {
